requirements.txt
django
django-storages
django-phonenumber-field
django-rest-framework
django-countries-plus
phonenumbers
whitenoise
africastalking
fcm-django
